Reduce code duplication and implement server placeholder on velocity
