+++ src/main/java/net/pistonmaster/pistonqueue/velocity/PistonQueueVelocity.java
@@ -344,7 +344,7 @@ public class PistonQueueVelocity {
         if (type.getDurationToPosition().containsKey(position)) {
             Duration duration = type.getDurationToPosition().get(position);
 
-            return ChatUtils.formatDuration(text, duration, position);
+            return SharedChatUtils.formatDuration(text, duration, position);
         } else {
             AtomicInteger biggestPositionAtomic = new AtomicInteger();
             AtomicReference<Duration> bestDurationAtomic = new AtomicReference<>(Duration.ZERO);
@@ -363,7 +363,7 @@ public class PistonQueueVelocity {
 
             Duration imaginaryDuration = biggestDuration.plus(difference, ChronoUnit.MINUTES);
 
-            return ChatUtils.formatDuration(text, imaginaryDuration, position);
+            return SharedChatUtils.formatDuration(text, imaginaryDuration, position);
         }
     }
